remove Link component and CSR logic in InlineLink
