+++ src/app/components/InlineLink/index.tsx
@@ -3,10 +3,12 @@
 import { useContext, FC, HTMLAttributes } from 'react';
 import { jsx, Theme } from '@emotion/react';
 import Url from 'url-parse';
+import { Link as ClientSideLink } from 'react-router-dom';
+import { pathToRegexp } from 'path-to-regexp';
 
 import { FontVariant, GelFontSize } from '../../models/types/theming';
 import { ServiceContext } from '../../contexts/ServiceContext';
-import Link from '../Link';
+import { articlePath } from '../../routes/utils/regex';
 import styles from './index.styles';
 
 interface Props extends HTMLAttributes<HTMLElement> {
@@ -14,8 +16,8 @@ interface Props extends HTMLAttributes<HTMLElement> {
   className?: string;
   fontVariant?: FontVariant;
   size?: GelFontSize;
-  to: string;
   text: string;
+  to: string;
 }
 
 const bbcDomains = [
@@ -27,6 +29,14 @@ const bbcDomains = [
   'www.bbcrussian.com',
   'bbcrussian.com',
 ];
+
+const allowedCsrPaths = articlePath; // currently, we only allow client side routing to article pages
+
+const parseLocation = (location: string) =>
+  pathToRegexp(allowedCsrPaths, [], {
+    start: false,
+    end: false,
+  }).exec(location);
 
 const InlineLink: FC<Props> = ({
   allowCSR = false,
@@ -40,22 +50,46 @@ const InlineLink: FC<Props> = ({
   const { externalLinkText } = useContext(ServiceContext);
   const { hostname } = new Url(to);
   const isExternalLink = !bbcDomains.some(bbcDomain => hostname === bbcDomain);
+  const linkProps = {
+    ...(isExternalLink &&
+      typeof text === 'string' && {
+        'aria-label': text.concat(externalLinkText),
+      }),
+    className,
+    css: ({ fontSizes, fontVariants }: Theme) => [
+      styles.self,
+      size && fontSizes[size],
+      fontVariant && fontVariants[fontVariant],
+    ],
+    ...htmlAttributes,
+  };
+
+  if (allowCSR) {
+    const parsedLocation = parseLocation(to);
+    const isCsr = parsedLocation?.length;
+
+    if (isCsr) {
+      const [pathname] = parsedLocation;
+      const [, hash] = to.split('#');
+
+      return (
+        <ClientSideLink
+          {...linkProps}
+          to={{
+            pathname,
+            hash,
+          }}
+        >
+          {text}
+        </ClientSideLink>
+      );
+    }
+  }
 
   return (
-    <Link
-      allowCSR={allowCSR}
-      className={className}
-      css={({ fontSizes, fontVariants }: Theme) => [
-        styles.self,
-        size && fontSizes[size],
-        fontVariant && fontVariants[fontVariant],
-      ]}
-      aria-label={isExternalLink ? text.concat(externalLinkText) : undefined}
-      to={to}
-      {...htmlAttributes}
-    >
+    <a {...linkProps} href={to}>
       {text}
-    </Link>
+    </a>
   );
 };
